requirements.txt
streamlit  
unifyai
git+https://github.com/unifyai/langchain.git#egg=langchain&subdirectory=libs/langchain
git+https://github.com/unifyai/langchain.git#egg=langchain-unify&subdirectory=libs/partners/unify
langchain_community 
youtube_search
duckduckgo-search
langchainhub
